ledger-on-memory: Split the Writer from the Reader.
There's no real reason to connect the two.

CHANGELOG_BEGIN
CHANGELOG_END
