+++ ledger/ledger-on-memory/src/main/scala/com/daml/ledger/on/memory/InMemoryLedgerReaderWriter.scala
@@ -3,18 +3,14 @@
 
 package com.daml.ledger.on.memory
 
-import akka.NotUsed
 import akka.stream.Materializer
-import akka.stream.scaladsl.Source
 import com.daml.api.util.TimeProvider
 import com.daml.caching.Cache
-import com.daml.dec.DirectExecutionContext
-import com.daml.ledger.api.health.{HealthStatus, Healthy}
 import com.daml.ledger.participant.state.kvutils.DamlKvutils.{DamlStateKey, DamlStateValue}
 import com.daml.ledger.participant.state.kvutils.api._
 import com.daml.ledger.participant.state.kvutils.export.LedgerDataExporter
 import com.daml.ledger.participant.state.kvutils.{Envelope, KeyValueCommitting, Raw}
-import com.daml.ledger.participant.state.v1.{LedgerId, Offset, ParticipantId, SubmissionResult}
+import com.daml.ledger.participant.state.v1.{LedgerId, ParticipantId}
 import com.daml.ledger.resources.{Resource, ResourceContext, ResourceOwner}
 import com.daml.ledger.validator.batch.{
   BatchedSubmissionValidator,
@@ -31,48 +27,16 @@ import com.daml.logging.LoggingContext.newLoggingContext
 import com.daml.metrics.Metrics
 import com.daml.platform.akkastreams.dispatcher.Dispatcher
 
-import scala.concurrent.{ExecutionContext, Future}
-import scala.util.Success
-
-final class InMemoryLedgerReaderWriter private[memory] (
-    override val ledgerId: LedgerId,
-    override val participantId: ParticipantId,
-    dispatcher: Dispatcher[Index],
-    state: InMemoryState,
-    validateAndCommit: ValidateAndCommit,
-    metrics: Metrics,
-) extends LedgerReader
-    with LedgerWriter {
-  override def commit(
-      correlationId: String,
-      envelope: Raw.Value,
-      metadata: CommitMetadata,
-  ): Future[SubmissionResult] =
-    validateAndCommit(correlationId, envelope, participantId)
-      .andThen { case Success(SubmissionResult.Acknowledged) =>
-        dispatcher.signalNewHead(state.newHeadSinceLastWrite())
-      }(DirectExecutionContext)
-
-  override def events(startExclusive: Option[Offset]): Source[LedgerRecord, NotUsed] =
-    reader.events(startExclusive)
-
-  override def currentHealth(): HealthStatus = Healthy
-
-  private val reader = new InMemoryLedgerReader(ledgerId, dispatcher, state, metrics)
-}
+import scala.concurrent.ExecutionContext
 
 object InMemoryLedgerReaderWriter {
-  val DefaultTimeProvider: TimeProvider = TimeProvider.UTC
-
-  type StateValueCache = Cache[DamlStateKey, DamlStateValue]
-
   final class BatchingOwner(
       ledgerId: LedgerId,
       batchingLedgerWriterConfig: BatchingLedgerWriterConfig,
       participantId: ParticipantId,
       metrics: Metrics,
-      timeProvider: TimeProvider = DefaultTimeProvider,
-      stateValueCache: StateValueCache = Cache.none,
+      timeProvider: TimeProvider = InMemoryLedgerWriter.DefaultTimeProvider,
+      stateValueCache: InMemoryLedgerWriter.StateValueCache = Cache.none,
       dispatcher: Dispatcher[Index],
       state: InMemoryState,
       engine: Engine,
@@ -91,31 +55,30 @@ object InMemoryLedgerReaderWriter {
           stateValueCache,
           ledgerDataExporter,
         )
-        readerWriter = new InMemoryLedgerReaderWriter(
-          ledgerId,
+        reader = new InMemoryLedgerReader(ledgerId, dispatcher, state, metrics)
+        writer = new InMemoryLedgerWriter(
           participantId,
           dispatcher,
           state,
           committer,
-          metrics,
         )
         // We need to generate batched submissions for the validator in order to improve throughput.
         // Hence, we have a BatchingLedgerWriter collect and forward batched submissions to the
         // in-memory committer.
-        ledgerWriter <- newLoggingContext { implicit loggingContext =>
+        batchingWriter <- newLoggingContext { implicit loggingContext =>
           ResourceOwner
-            .forCloseable(() => BatchingLedgerWriter(batchingLedgerWriterConfig, readerWriter))
+            .forCloseable(() => BatchingLedgerWriter(batchingLedgerWriterConfig, writer))
             .acquire()
         }
-      } yield createKeyValueLedger(readerWriter, ledgerWriter)
+      } yield createKeyValueLedger(reader, batchingWriter)
   }
 
   final class SingleParticipantBatchingOwner(
       ledgerId: LedgerId,
       batchingLedgerWriterConfig: BatchingLedgerWriterConfig,
       participantId: ParticipantId,
-      timeProvider: TimeProvider = DefaultTimeProvider,
-      stateValueCache: StateValueCache = Cache.none,
+      timeProvider: TimeProvider = InMemoryLedgerWriter.DefaultTimeProvider,
+      stateValueCache: InMemoryLedgerWriter.StateValueCache = Cache.none,
       metrics: Metrics,
       engine: Engine,
   )(implicit materializer: Materializer)
@@ -145,17 +108,15 @@ object InMemoryLedgerReaderWriter {
       participantId: ParticipantId,
       keySerializationStrategy: StateKeySerializationStrategy,
       metrics: Metrics,
-      timeProvider: TimeProvider = DefaultTimeProvider,
-      stateValueCache: StateValueCache = Cache.none,
+      timeProvider: TimeProvider = InMemoryLedgerWriter.DefaultTimeProvider,
+      stateValueCache: InMemoryLedgerWriter.StateValueCache = Cache.none,
       dispatcher: Dispatcher[Index],
       state: InMemoryState,
       engine: Engine,
   )(implicit materializer: Materializer)
       extends ResourceOwner[KeyValueLedger] {
     override def acquire()(implicit context: ResourceContext): Resource[KeyValueLedger] = {
-      val keyValueCommitting =
-        createKeyValueCommitting(metrics, timeProvider, engine)
-
+      val keyValueCommitting = createKeyValueCommitting(metrics, timeProvider, engine)
       val committer = createPreExecutingCommitter(
         keyValueCommitting,
         keySerializationStrategy,
@@ -164,17 +125,14 @@ object InMemoryLedgerReaderWriter {
         timeProvider,
         stateValueCache,
       )
-
-      val readerWriter = new InMemoryLedgerReaderWriter(
-        ledgerId,
+      val reader = new InMemoryLedgerReader(ledgerId, dispatcher, state, metrics)
+      val writer = new InMemoryLedgerWriter(
         participantId,
         dispatcher,
         state,
         committer,
-        metrics,
       )
-
-      Resource.successful(createKeyValueLedger(readerWriter, readerWriter))
+      Resource.successful(createKeyValueLedger(reader, writer))
     }
   }
 
@@ -184,7 +142,7 @@ object InMemoryLedgerReaderWriter {
       state: InMemoryState,
       metrics: Metrics,
       timeProvider: TimeProvider,
-      stateValueCache: StateValueCache,
+      stateValueCache: InMemoryLedgerWriter.StateValueCache,
       ledgerDataExporter: LedgerDataExporter,
   )(implicit materializer: Materializer): ValidateAndCommit = {
     val validator = BatchedSubmissionValidator[Index](
@@ -228,7 +186,7 @@ object InMemoryLedgerReaderWriter {
       state: InMemoryState,
       metrics: Metrics,
       timeProvider: TimeProvider,
-      stateValueCache: StateValueCache,
+      stateValueCache: InMemoryLedgerWriter.StateValueCache,
   )(implicit materializer: Materializer): ValidateAndCommit = {
     val committer = new PreExecutingValidatingCommitter[
       Option[DamlStateValue],
